Refactor: Migrate to tab-based systems check with color system tests
