+++ REFACTORING_RESTORATION_PLAN.md
@@ -3,10 +3,11 @@
 ## 📌 Important Reminders for Agent
 
 1. **After EVERY phase completion**: Update Task 6.1 (Systems Check) to include new functionality demonstrations
-2. **Clean up this plan**: Remove completed phases from this document to reduce token usage
-3. **Track progress**: Update `completed_REFACTORING_RESTORATION_PLAN.md` with implementation details
-4. **Success metrics**: Are tracked in the completed tracker file, not here
-5. **⚠️ NEW: Create completed documents**: Starting from Phase 2, create new `completed_PHASE_X_REFACTORING.md` documents for each phase instead of just updating the main completed file
+2. **⚠️ CRITICAL: Incorporate each phase into SystemsCheck**: After completing any phase, add tests/demos for that phase's functionality to the systems check interface
+3. **Clean up this plan**: Remove completed phases from this document to reduce token usage
+4. **Track progress**: Update `completed_REFACTORING_RESTORATION_PLAN.md` with implementation details
+5. **Success metrics**: Are tracked in the completed tracker file, not here
+6. **⚠️ NEW: Create completed documents**: Starting from Phase 2, create new `completed_PHASE_X_REFACTORING.md` documents for each phase instead of just updating the main completed file
 
 ## Overview
