+++ completed_PHASE_2_REFACTORING.md
@@ -71,6 +71,12 @@ This document tracks Phase 2 implementation progress: fixing component integrati
    - Old 596-line `useColorSystem.ts` deleted
    - All import references updated
 
+5. **Systems Check Integration**: 
+   - Replaced floating button/modal with tabs-based interface
+   - Added Phase 2 color system tests to SystemsCheck component
+   - Interactive color generation and glassmorphism testing
+   - Component migration status display
+
 ---
 
 ## Success Criteria for Phase 2
